+++ src/Query.ts
@@ -93,6 +93,12 @@ export class Query<T = unknown> implements PromiseLike<T> {
     }
   }
 
+  clearData = (): void => {
+    action(() => {
+      this.data = undefined
+    })
+  }
+
   refetch = (): Promise<T> => {
     return Promise.resolve().then(
       action(() => {
@@ -128,7 +134,7 @@ export class Query<T = unknown> implements PromiseLike<T> {
         this.error = false
         this.data = data
       }),
-      action(error => {
+      action((error) => {
         this.loading = false
         this.error = error
       })
@@ -164,10 +170,10 @@ export class Query<T = unknown> implements PromiseLike<T> {
   ): PromiseLike<TResult1 | TResult2>
   then(onfulfilled: any, onrejected: any) {
     return this.promise.then(
-      d => {
+      (d) => {
         this.store.__runInStoreContext(() => onfulfilled(d))
       },
-      e => {
+      (e) => {
         this.store.__runInStoreContext(() => onrejected(e))
       }
     )